backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
pandas>=2.0.0
reportlab>=4.0.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
sendgrid>=6.9.0
certifi>=2023.0.0
gunicorn>=21.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.0
